ADMIN LOGGIN

+++ server/api/users.js
@@ -52,12 +52,15 @@ router.post('/:userId/mycart', async (req, res, next) => {
 router.delete('/:cartItemId/mycart', async (req, res, next) => {
   try {
     const cartItemId = req.params.cartItemId
-    const response = await CartItem.destroy({
-      where: {
-        id: cartItemId
-      }
-    })
-    res.json(response)
+
+    if (cartItemId !== 'undefined') {
+      const response = await CartItem.destroy({
+        where: {
+          id: cartItemId
+        }
+      })
+      res.json(response)
+    }
   } catch (error) {
     next(error)
   }
